add request timeout

+++ config/config.go
@@ -41,7 +41,7 @@ type RendoraConfig struct {
 		URL         string   `valid:"requrl"`
 		AuthToken   string   `mapstructure:"authToken"`
 		BlockedURLs []string `mapstructure:"blockedURLs"`
-		Timeout     uint16   `valid:"range(5|30)"`
+		Timeout     int64    `valid:"range(5|30)"`
 		Internal    struct {
 			URL string `valid:"url"`
 		}
